Add WidevineCdm adapter dir to winstaller
Auditors: @bsclifton

This doesn't actually include the CDM, but it includes a supporting file that is expecting to be installed. 0.13.0 uses a different location for this file than 0.12.15 used so we need to preserve this new directory and we need to install it

When the component updater installs widevine (if the user explicitly
does so), the adapter file is copied to the needed location from the
install path.  It used to do this as well it just used to be in a
different location.